Add tests for subsscribeToRule.
Changes to .sol:
1. Better error messages
2. Fixed Bug with  addRule where the constraints werent actually being saved.
3. subscribeToRule wasnt marked as payable so we couldnt send it ETH.

Also added a couple of tests to addRule to check if multiple rules get added properly. Currently the tests where rules with same params result in same ruleHash is failing because the block timestamp is different. Hardhat automines a block with every tx, so that needs to be adjusted for this rule to pass.