Replace UNI with TST1
We've been using TST1 as a UNI replacement in the code.
We initially called it UNI to make it easier to think through the flow when price was involved.
This is misleading in the code, so code should not call it UNI anymore.
Furthermore, it may imply a link between what token we are using and swapUniSingleAction to future readers - which is also misleading.