Fix compile errors. Utils fns need to be mark internal so it gets bundled with our contract. If it's called public, the our contract needs a reference to the library which is why the deployment was failing.
Also fixed a bug where in some cases calldata (where the action is passed in from the user), and in other cases memory (where the Action is generated using abi.decode) was used to create the hash. There doent seem to be way for a fn to accept either calldata / memory and forward it to abi.encode so we call ebi.encode directly and remove the extra fn call.